Created choropleth weather map Improved search functionality Modularized and cleaned up code
