+++ HTML/weathertutorial.html
@@ -108,96 +108,10 @@
   </nav>
   <!--navbar ends here-->
 
-  <div class="content">
-    <div>
-    <div class="search-bar d-flex w-50 mx-auto mb-2 my-2">
-      <input class="form-control me-2" id="search-input" type="text" placeholder="Enter a city or country">
-      <button class="btn btn-secondary btn-outline-light" id="search-button" type="button">Search</button>
-    </div>
-    <ul class="w-50 mx-auto" id="suggestions"></ul>
-  </div>
-
-  <div class="container d-flex m-auto" id="container">
-    <div class="aside container" id="aside">
-
-        <!-- <h1>Air Quality</h1> -->
-        <!--Severe Weather Alert-->
-    </div>
-    <!--tab nav options-->
-    <div class="container tabs-container m-auto" id="tabs">
-      <div class="tabs d-flex mb-2">
-        <a class="tab" id="tab1" data-tab="1" role="tab">Today</a>
-        <a class="tab" id="tab2" data-tab="2" role="tab">Tomorrow</a>
-        <a class="tab" id="tab3" data-tab="3" role="tab">7 days</a>
-        <!-- <a class="tab" id="tab4" data-tab="4" role="tab">Monthly</a> -->
-        <a class="tab" id="tab5" data-tab="5" role="tab">Visualization</a>
-      </div>
-
-    <!--Tab contents-->
-    <div class="tab-content" id="tab-content">
-      <!--Tab1 content-->
-      <div class="tabcontent fade show" id="tabcontent1" data-tab="1" role="tabpane">
-        <!--Tab1 header dynamically added with javascript-->
-        <!--Tab1 body dynamically added with javascript-->
-        <!--Tab1 daytime section dynamically added with javascript-->
-        <!--hour section dynamically added with javascript-->
-      </div>
-      <!--Tab2 content-->
-      <div class="tabcontent fade show" id="tabcontent2" data-tab="2" role="tabpane">
-      </div>
-      <!--Tab3 content-->
-      <div class="tabcontent fade show" id="tabcontent3" data-tab="3" role="tabpane">
-      </div>
-      <!--Tab4 content-->
-      <div class="tabcontent fade show" id="tabcontent5" data-tab="4" role="tabpane">
-        <h3 id="title">Weather Heatmap</h1>
-          <div class="heatmap-container">
-        <svg id="heatmap-svg" viewBox="0 0 1200 600"></svg>
-        <div id="tooltip"></div>
-      </div>
-        <div id="map"></div>
-        <!--Temperature legends-->
-        <div id="map-legend">
-          <div class="legend-item">
-            <span class="color" style="background-color: blue;"></span>
-            <span class="label">-10°C to 0°C</span>
-          </div>
-          <div class="legend-item">
-            <span class="color" style="background-color: cyan;"></span>
-            <span class="label">1°C to 10°C</span>
-          </div>
-          <div class="legend-item">
-            <span class="color" style="background-color: green;"></span>
-            <span class="label">11°C to 20°C</span>
-          </div>
-          <div class="legend-item">
-            <span class="color" style="background-color: yellow;"></span>
-            <span class="label">21°C to 30°C</span>
-          </div>
-          <div class="legend-item">
-            <span class="color" style="background-color: orange;"></span>
-            <span class="label">31°C to 40°C</span>
-          </div>
-          <div class="legend-item">
-            <span class="color" style="background-color: red;"></span>
-            <span class="label">41°C to 100°C</span>
-          </div>
-        </div>
-        <div id="weather-details"></div>
-      </div>
-    </div>
-    </div>
-
-  <!--Dark mode-->
-    <div class="dark-mode-switch">
-      <label class="switch-label" for="dark-mode-switch">Dark mode</label>
-      <label class="switch">
-        <input type="checkbox" id="dark-mode-switch">
-        <span class="slider round"></span>
-      </label>
-    </div>
-  </div>
-</div>
+    <!--Main content-->
+    <section id="tutorial">
+        
+    </section>
 
 
 
@@ -273,15 +187,11 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz"
     crossorigin="anonymous"></script>
-
-    <!-- <script src="../JS/weathermap4.js"></script> -->
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDQb008yhHF8gs50DFD_67HaJopYtFfa2Y&callback=initMap&libraries=&v=weekly" async defer></script>
+    <!-- <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDQb008yhHF8gs50DFD_67HaJopYtFfa2Y&libraries=visualization&callback=initMap"></script> -->
 
   <!--External Javascript file-->
-  <!-- <script type="module" src="../JS/weatherviz.js"></script> -->
-  <!-- <script type="module" src="../JS/weekly_weather.js"></script>
-  <script type="module" src="../JS/weatherUtils.js"></script> -->
-  <script type="module" src="../JS/weather.js"></script>
+  <script src="../JS/weather.js"></script>
+  <!-- <script type="module" src="../JS/weekly_weather.js"></script> -->
 
 </body>
