--- JS/severe_weather.js
@@ -1,66 +1,0 @@
-document.addEventListener("DOMContentLoaded", function () {
-    // DOM Elements
-    const tabs = document.querySelectorAll('.tab')
-    const tabContents = document.querySelectorAll('.tabcontent')
-    const darkModeSwitch = document.querySelector('#dark-mode-switch')
-  
-    // Functions
-    const activateTab = tabnum => {
-  
-      tabs.forEach(tab => {
-        tab.classList.remove('active')
-      })
-  
-      tabContents.forEach(tabContent => {
-        tabContent.classList.remove('active')
-      })
-  
-      document.querySelector('#tab' + tabnum).classList.add('active')
-      document.querySelector('#tabcontent' + tabnum).classList.add('active')
-      localStorage.setItem('jstabs-opentab', JSON.stringify(tabnum))
-  
-    }
-  
-    // Event Listeners
-    tabs.forEach(tab => {
-      tab.addEventListener('click', () => {
-        activateTab(tab.dataset.tab)
-      })
-    })
-  
-    darkModeSwitch.addEventListener('change', () => {
-      document.querySelector('body').classList.toggle('darkmode')
-      localStorage.setItem('jstabs-darkmode', JSON.stringify(!darkmodeModeSwitch.checked));
-    })
-  
-    // Retrieve stored data
-    let darkmode = JSON.parse(localStorage.getItem('jstabs-darkmode'))
-    const opentab = JSON.parse(localStorage.getItem('jstabs-opentab')) || '3'
-  
-    // and..... Action!
-    if (darkmode === null) {
-      darkmode = window.matchMedia("(prefers-color-scheme: dark)").matches // match to OS theme
-    }
-    if (darkmode) {
-      document.querySelector('body').classList.add('darkmode')
-      document.querySelector('#dark-mode-switch').checked = 'true'
-    }
-    activateTab(opentab)
-
-
-
-});
-
-   // const newsDiv = createDiv('news');
-                    // const newsTitle = createDivWithLabel('news-title', 'Title', newsdata.title);
-                    // const newsDescription = createDivWithLabel('news-description', 'Description', newsdata.description);
-                    // const newsSource = createDivWithLabel('news-source', 'Source', newsdata.source);
-                    // const newsUrl = createDivWithLabel('news-url', 'URL', newsdata.url);
-                    // const newsPublishedAt = createDivWithLabel('news-published-at', 'Published At', newsdata.publishedAt);
-                    // const newsAuthor = createDivWithLabel('news-author', 'Author', newsdata.author);
-                    // const newsImage = createDivWithLabel('news-image', 'Image', newsdata.image);
-                    // const newsContent = createDivWithLabel('news-content', 'Content', newsdata.content);
-                    // const newsCategory = createDivWithLabel('news-category', 'Category', newsdata.category);
-                    // const newsLanguage = createDivWithLabel('news-language', 'Language', newsdata.language);
-                    // const newsCountry = createDivWithLabel('news-country', 'Country', newsdata.country);
-
